refactor main layout into modular components
